Removed `ListAssessmentResults` from assessment (#979)
* Removed `ListAssessmentResults` from assessment

This finally removes the `ListAssessmentResults` from the assessment service, which caused confusion. This also moves the `ListAssessmentResultsRequest` and response to the `orchestrator` package.

---------
