+++ api/assessment/assessment.proto
@@ -60,42 +60,7 @@ service Assessment {
   // stream. Part of the public API. Not exposed as REST.
   rpc AssessEvidences(stream AssessEvidenceRequest)
       returns (stream AssessEvidencesResponse) {};
-
-  // List all assessment results of all cloud services the requester can access.
-  // Optionally, further filters of specific cloud services, metrics or the
-  // compliance status can be specified.
-  //
-  // Part of the public API, also exposed as REST.
-  rpc ListAssessmentResults(ListAssessmentResultsRequest)
-      returns (ListAssessmentResultsResponse) {
-    option (google.api.http) = {
-      get : "/v1/assessment/results"
-    };
-  }
 };
-
-message ListAssessmentResultsRequest {
-  // Optional. List only assessment results of a specific cloud service.
-  optional string filtered_cloud_service_id = 1
-      [ (validate.rules).string.uuid = true ];
-  // Optional. List only compliant assessment results.
-  optional bool filtered_compliant = 2;
-  // Optional. List only assessment results of a specific metric id.
-  repeated string filtered_metric_id = 3
-      [ (validate.rules).repeated .items.string.min_len = 1 ];
-
-  // Optional. Latest results grouped by resource_id and metric_id.
-  optional bool latest_by_resource_id = 4;
-
-  int32 page_size = 10;
-  string page_token = 11;
-  string order_by = 12;
-  bool asc = 13;
-}
-message ListAssessmentResultsResponse {
-  repeated AssessmentResult results = 1;
-  string next_page_token = 2;
-}
 
 message ConfigureAssessmentRequest {}
 message ConfigureAssessmentResponse {}
