+++ api/assessment/assessment_test.go
@@ -29,7 +29,6 @@ import (
 	"testing"
 
 	"clouditor.io/clouditor/internal/testdata"
-	"clouditor.io/clouditor/internal/util"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
@@ -272,59 +271,6 @@ func Test_ValidateAssessmentResult(t *testing.T) {
 			} else {
 				assert.NoError(t, err)
 			}
-		})
-	}
-}
-
-func TestListAssessmentResultsRequest_Validate(t *testing.T) {
-	type fields struct {
-		req *ListAssessmentResultsRequest
-	}
-
-	tests := []struct {
-		name    string
-		fields  fields
-		wantErr assert.ErrorAssertionFunc
-	}{
-		{
-			name: "Request is empty",
-			fields: fields{
-				&ListAssessmentResultsRequest{},
-			},
-			wantErr: assert.NoError,
-		},
-		{
-			name: "Invalid cloud service id",
-			fields: fields{
-				req: &ListAssessmentResultsRequest{
-					FilteredCloudServiceId: util.Ref("invalidCloudServiceId"),
-				},
-			},
-			wantErr: func(tt assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorContains(t, err, "FilteredCloudServiceId: value must be a valid UUID")
-			},
-		},
-		{
-			name: "No filtered cloud service id",
-			fields: fields{
-				req: &ListAssessmentResultsRequest{},
-			},
-			wantErr: assert.NoError,
-		},
-		{
-			name: "Happy path",
-			fields: fields{
-				req: &ListAssessmentResultsRequest{
-					FilteredCloudServiceId: util.Ref(testdata.MockCloudServiceID),
-				},
-			},
-			wantErr: assert.NoError,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := tt.fields.req.Validate()
-			tt.wantErr(t, err)
 		})
 	}
 }
